\name{HalloWelt}
\alias{HalloWelt}
\title{Hallo Welt}
\usage{
  HalloWelt(pkg = "plyr", repo = "http://cran.r-mirror.de")
}
\arguments{
  \item{pkg}{package wanted}

  \item{repo}{where to get it}
}
\description{
  Say hello and install my package, please
}
\examples{
HalloWelt("plyr", "http://cran.r-mirror.de")
HalloWelt("plutils", "http://R-Forge.R-project.org")
HalloWelt("fdnonpar", "http://R-Forge.R-project.org")
}

