\name{hasType}
\alias{currentType}
\alias{hasType}
\title{Manipulate s.o.s. type information}
\usage{
  hasType(x, type = .TYPES)

  currentType(x)
}
\arguments{
  \item{x}{point pattern, of class sostyppp}

  \item{type}{character, type of second-order stationarity.
  See \code{\link{sostyppp}} for details.}
}
\description{
  Access the encrypted type information of second-order
  stationarity typed point patterns, objects of class
  \code{sostyppp}.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}

