\name{identxy}
\alias{identxy}
\title{identical mapping}
\usage{
  identxy(x, y = NULL)
}
\arguments{
  \item{x}{a numerical vector, or list or dataframe with
  elements \code{x} and \code{y}}

  \item{y}{optional, needed if \code{x} is a vector only}
}
\description{
  identical mapping
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}
\keyword{manip}
\keyword{spatial}

