\name{is.ppsample}
\alias{is.ppsample}
\title{Check whether an object is a sample of point patterns}
\usage{
  is.ppsample(x)
}
\arguments{
  \item{x}{any \code{R} object}
}
\value{
  \code{TRUE} if \code{x} belongs to class
  \code{"ppsample"}, otherwise \code{FALSE}.
}
\description{
  Checks if an object belongs to class \code{"ppsample"}.
}

