\name{is.sostpp}
\alias{is.sostpp}
\title{Check whether an object is a second-order stationarity typed point pattern}
\usage{
  is.sostpp(x)
}
\arguments{
  \item{x}{any \code{R} object}
}
\value{
  \code{TRUE} if \code{x} belongs to class \code{"sostpp"},
  otherwise \code{FALSE}.
}
\description{
  Checks if an object belongs to class \code{"sostpp"}.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}
\seealso{
  \code{\link{sostpp.object}} for details on the class.
}

