% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sostyppp.R
\name{is.sostyppp}
\alias{is.sostyppp}
\title{Check whether an object is a second-order stationarity typed point pattern}
\usage{
is.sostyppp(x)
}
\arguments{
\item{x}{any \code{R} object}
}
\value{
\code{TRUE} if \code{x} belongs to class \code{"sostyppp"}, otherwise \code{FALSE}.
}
\description{
Checks if an object belongs to class \code{"sostyppp"}.
}
\author{
Ute Hahn,  \email{ute@imf.au.dk}
}
\seealso{
\code{\link{sostyppp}} for details on the class.
}

