\name{print.sostyppp}
\alias{print.sostyppp}
\title{Print brief details of a second-order stationarity typed point pattern}
\usage{
  \method{print}{sostyppp} (x, ...)
}
\arguments{
  \item{x}{sos-typed point pattern.}

  \item{...}{ignored}
}
\description{
  Extends spatstat method \code{\link{print.ppp}} by type
  of second-order stationarity.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}
\seealso{
  \code{\link{print.ppp}} for the print method of class
  ancestor \code{ppp}.
}

