\name{sostatpp-package}
\alias{sostatpp-package}
\alias{sostatpp}
\docType{package}
\title{
Try to reinstall plutils and fdnonpar
~~ sostatpp dummy version ~~
}
\description{
More about what it does (maybe more than one line)
Not really much
}
\details{
\tabular{ll}{
Package: \tab sostatppdummy\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-12\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ute Hahn

Maintainer: Ute Hahn <ute@imf.au.dk>
}
\keyword{ package }
\examples{
helloworld("keep fingers crossed")
}
