% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sostatppoptions.R
\name{sostatpp.options}
\alias{sostatpp.options}
\title{Options for package sostatpp}
\usage{
sostatpp.options(...)
}
\arguments{
\item{...}{option names given as character string, or given as \code{name = value}.
Alternatively, options can be passed as \code{\link{simplist}}s.}
}
\value{
For \code{ploptions()}, a \code{simplist} of set options for package \code{plutils};
if invoked with names of options, a list of all options contained in the arguments.
If called with \code{name = value}, no visible result is returned.
}
\description{
Get or set options for package sostatpp
}
\details{
It is possible to set any options - this might be useful for global options of
other packages. This can however also be done with the base function \code{options}, the
only difference is that \code{sostatppoptions} returns a \code{\link{simplist}}, and resolves
\code{\link{simplist}} arguments, but not \code{\link{list}} arguments.

Currently, only one options is used by package \code{sostatpp}:
\tabular{ll}{
\code{Kstyles}\tab list with elements \code{theo}, \code{iso}, \code{trans} and
\code{border}, that control plotting of $K$-functions obtained by \code{\link{estK}},
or other \code{funsamples} containing elements \code{theo}, \ldots, \code{border}.
By default, each element of \code{Kstyles} is a \code{simplist} with entries \code{col} and
\code{lty}.
}
}
\seealso{
The base function \code{\link{options}} is similar.
}

