\name{twoquadsets}
\alias{twoquadsets}
\title{Set up samples of quadrats with high and low intensity}
\usage{
  twoquadsets(pp, quads = NULL, nx = NULL, ny = NULL,
    gradx = FALSE, grady = FALSE, xbreaks = NULL,
    ybreaks = NULL, minpoints = 20,
    by.intensity = !xor(gradx, grady))
}
\arguments{
  \item{pp}{The point pattern, object of class \code{ppp}
  or \code{sostyppp}}

  \item{quads}{optional, a tessellation (class \code{tess})
  or a list of windows (class \code{owin}) that is to be
  divided into subsets}

  \item{nx,ny}{number of quadrats in the \code{x} and
  \code{y} directions}

  \item{gradx,grady}{logical, if \code{TRUE}, the \eqn{x}
  resp. \eqn{y} coordinates of the pattern \code{pp} are
  used to determine the boundaries}

  \item{xbreaks,ybreaks}{numeric vector of \eqn{x} resp.
  \eqn{y} coordinates of the quadrat boundaries}

  \item{minpoints}{minimum number of points required, see
  Details.}

  \item{by.intensity}{logical, if \code{TRUE}, use
  estimated intensity for division, otherwise, go by
  gradient coordinate.}
}
\value{
  A list with three elements \code{hi}, \code{lo} and
  \code{unused}. The elements themselves are lists of
  \code{spatstat::\link[spatstat]{owin}} objects.
}
\description{
  Divide a set of quadrats into sets with high and low
  estimated intensity of a point pattern. The quadrats can
  either be given, or generated by splitting the window of
  the point pattern into rectangular quadrats.
}
\details{
  If \code{quadrats} is not given, the quadrats are
  generated using spatstat's function
  \code{\link[spatstat]{quadrats}}, which splits the window
  into rectangles. Then, the \eqn{x}-coordinates of quadrat
  boundaries are determined either by specifying
  \code{xbreaks}, or by \code{nx} and \code{gradx}, with
  priority on \code{xbreaks}. If \code{gradx} is
  \code{TRUE}, equidistant quantiles of the
  \eqn{x}-coordinates of the point pattern \code{pp} are
  used; otherwise, the boundaries are equally spaced. The
  \eqn{y}-coordinate is treated analogously.

  Quadrats with a number of points less than
  \code{minpoints} are returned as a list "\code{unused}".

  If \code{by.intensity} is \code{TRUE}, the remaining
  quadrats are divided into a set (list of quadrats) with
  high and a set with low estimated intensity. #'If
  \code{by.intensity} is \code{FALSE}, the remaining
  quadrats are divided into two sets according to their
  \eqn{x}-coordinate, if \code{gradx} is \code{TRUE}.
  Otherwise, the \eqn{y}-coordinate is used as criterion.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}
\seealso{
  The spatstat function
  \code{spatstat::\link[spatstat]{quadrats}}
}

