\name{has.type}
\alias{currenttype}
\alias{has.type}
\title{Manipulate s.o.s. type information}
\usage{
  has.type(x, type = .TYPES)

  currenttype(x)
}
\arguments{
  \item{x}{point pattern, of class sostyppp}

  \item{type}{character, type of second-order stationarity.
  See \code{\link{sostyppp.object}} for details.}
}
\description{
  Access the encrypted type information of second-order
  stationarity typed point patterns, objects of class
  \code{sostyppp}.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}

