% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sostyppp.R
\name{print.sostyppp}
\alias{print.sostyppp}
\title{Print brief details of a second-order stationarity typed point pattern}
\usage{
\method{print}{sostyppp}(x, ...)
}
\arguments{
\item{x}{sos-typed point pattern.}

\item{...}{ignored}
}
\description{
Extends spatstat method \code{\link[spatstat]{print.ppp}} by type of second-order stationarity.
}
\author{
Ute Hahn,  \email{ute@imf.au.dk}
}
\seealso{
\code{\link{print.ppp}} for the print method of class ancestor \code{ppp}.
}

