% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kest.R, R/MakeHidden.R, R/edgecrossRipley.R, R/lsdists.R, R/sostyppp.R, R/transformxy-generic.R
\name{ls.r.args}
\alias{.settype}
\alias{currentType<-}
\alias{edgecross.Ripley}
\alias{getIntensity}
\alias{longTypeName}
\alias{ls.r.args}
\alias{lsclosepairs}
\alias{lspairdist}
\alias{mapstructxy}
\alias{normalizedIntensity}
\alias{peripolyxy}
\alias{sostatpp-internal}
\alias{subdivpolyxy}
\title{Internal functions of sostatpp package}
\usage{
ls.r.args(r = NULL, eps = NULL, rmaxdefault = 4, minscale = NULL,
  maxscale)

getIntensity(X, intensity = NULL, ...)

normalizedIntensity(X, intensity = NULL, normpower = 2)

edgecross.Ripley(X, r, dphi = pi/4, W = X$window, maxweight = 100)

lspairdist(x, y, invscale, lambda)

lsclosepairs(X, rmax, invscale, lambda, ordered = TRUE, what = c("all",
  "indices"))

currentType(x) <- value

longTypeName(type)

.settype(type, typecode)

mapstructxy(X, mapxy = identxy, ...)

peripolyxy(poly)

subdivpolyxy(poly, newlen)
}
\description{
Functions for use by the package's functions, partly slightly
documented in the source code
}
\keyword{internal}

