\name{Subset_ppsample}
\alias{Subset_ppsample}
\alias{[.ppsample}
\title{Extract subset of a point pattern sample}
\usage{
  \method{[}{ppsample} (x, i)
}
\arguments{
  \item{x}{a list of point patterns, object of class
  \code{"ppsample"}.}

  \item{i}{subset index.}
}
\value{
  A \code{ppsample} object.
}
\description{
  Subsample a point pattern sample, retaining information
  about the original pattern's window.
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}

